BaseFile/BaseDirectory
I created base classes for the IDirectory and IFile with the basic logic expected from the library.
Any actual "File" operations are done in the concrete classes, such as the PhysicalDirectory and PhysicalFile.

The PhysicalDirectory and PhysicalFile already implement the new base types, and execute only handle the actual creation, deletion, etc of the file.
Anything else, like checking if the Directory of the file exists, and creating it, before creating the file, is done in the BaseClass.